+++ src/components/TodoCount.js
@@ -1,6 +1,6 @@
 import TodoModel from "../model/todoModel.js";
-import { todoCountTemplate } from "../../utils/template.js";
-import { SELECTOR } from "../../utils/constants.js";
+import { todoCountTemplate } from "../utils/template.js";
+import { SELECTOR } from "../utils/constants.js";
 
 export default function TodoCount() {
   if (new.target !== TodoCount) {
